Move HipFileDataOps to util.
